requirements.txt
altair==5.5.0
attrs==25.3.0
biopython==1.85
blinker==1.9.0
cachetools==5.5.2
cattrs==24.1.3
certifi==2025.4.26
charset-normalizer==3.4.2
chembl-webresource-client==0.10.9
click==8.2.0
colorama==0.4.6
easydict==1.13
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.39.0
numpy==2.2.5
packaging==24.2
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.30.2
py3Dmol==2.4.2
pyarrow==20.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
requests-cache==1.2.1
rpds-py==0.24.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.2
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
watchdog==6.0.0
